+++ crates/objc2/src/runtime/nsobject.rs
@@ -2,11 +2,8 @@ use core::fmt;
 use core::hash;
 
 use crate::rc::{DefaultId, Id, Owned};
-use crate::runtime::{Class, Object, Protocol};
-use crate::{
-    extern_methods, msg_send, msg_send_id, ClassType, ImplementedBy, Message, ProtocolObject,
-    ProtocolType,
-};
+use crate::runtime::{Class, ImplementedBy, Object, Protocol, ProtocolObject};
+use crate::{extern_methods, msg_send, msg_send_id, ClassType, Message, ProtocolType};
 
 crate::__emit_struct! {
     (